refactor: rename layout_handler strategies folder to types
