+++ src/ingestion_pipeline/chunking/layout_handler/types/table/line_chunker.py
@@ -8,8 +8,8 @@ from textractor.data.constants import LINE
 from textractor.entities.bbox import BoundingBox
 from textractor.entities.layout import Layout, Line
 
-from ingestion_pipeline.chunking.layout_handler.strategies.table.base import BaseTableChunker
-from ingestion_pipeline.chunking.layout_handler.strategies.table.schemas import TextBlock
+from ingestion_pipeline.chunking.layout_handler.types.table.base import BaseTableChunker
+from ingestion_pipeline.chunking.layout_handler.types.table.schemas import TextBlock
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
 
 logger = logging.getLogger(__name__)
