+++ tests/chunking/strategies/table/test_layout_table.py
@@ -8,9 +8,9 @@ from textractor.entities.table import Table
 
 from ingestion_pipeline.chunking.exceptions import ChunkException
 from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
-from ingestion_pipeline.chunking.layout_handler.strategies.table import LayoutTableChunkingStrategy
-from ingestion_pipeline.chunking.layout_handler.strategies.table.cell_chunker import CellTableChunker
-from ingestion_pipeline.chunking.layout_handler.strategies.table.line_chunker import LineTableChunker
+from ingestion_pipeline.chunking.layout_handler.types.table import LayoutTableChunkingStrategy
+from ingestion_pipeline.chunking.layout_handler.types.table.cell_chunker import CellTableChunker
+from ingestion_pipeline.chunking.layout_handler.types.table.line_chunker import LineTableChunker
 
 
 @pytest.fixture
